requirements.txt
torch
torchvision
tqdm
numpy
fastapi
uvicorn
python-multipart
streamlit
pillow
timm
grad-cam
opencv-python
